code cleanup
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@1176 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/Log_p.java
@@ -68,7 +68,7 @@ public class Log_p {
                 reversed=true;
             }
             if(post.containsKey("lines")){
-                lines = (int)Integer.parseInt((String)post.get("lines"));
+                lines = Integer.parseInt((String)post.get("lines"));
             }
         }
